Fixed that certificate template background images are saved as directories

+++ Services/Certificate/classes/Template/Action/Clone/class.ilCertificateCloneAction.php
@@ -120,11 +120,16 @@ class ilCertificateCloneAction
             $backgroundImageFile = basename($backgroundImagePath);
             $backgroundImageThumbnail = dirname($backgroundImagePath) . '/background.jpg.thumb.jpg';
 
-            $newBackgroundImage = $certificatePath . $backgroundImageFile;
-            $newBackgroundImageThumbnail = str_replace($webDir, '', $this->getBackgroundImageThumbPath($certificatePath));
-
-            if ($this->fileSystem->has($backgroundImagePath)) {
-                if ($this->fileSystem->has($newBackgroundImage)) {
+            $newBackgroundImage = '';
+            $newBackgroundImageThumbnail = '';
+            if ($this->fileSystem->has($backgroundImagePath) &&
+                !$this->fileSystem->hasDir($backgroundImagePath)
+            ) {
+                $newBackgroundImage = $certificatePath . $backgroundImageFile;
+                $newBackgroundImageThumbnail = str_replace($webDir, '', $this->getBackgroundImageThumbPath($certificatePath));
+                if ($this->fileSystem->has($newBackgroundImage) &&
+                    !$this->fileSystem->hasDir($newBackgroundImage)
+                ) {
                     $this->fileSystem->delete($newBackgroundImage);
                 }
 
@@ -134,8 +139,12 @@ class ilCertificateCloneAction
                 );
             }
 
-            if ($this->fileSystem->has($backgroundImageThumbnail)) {
-                if ($this->fileSystem->has($newBackgroundImageThumbnail)) {
+            if ($this->fileSystem->has($backgroundImageThumbnail) &&
+                !$this->fileSystem->hasDir($backgroundImageThumbnail)
+            ) {
+                if ($this->fileSystem->has($newBackgroundImageThumbnail) &&
+                    !$this->fileSystem->hasDir($newBackgroundImageThumbnail)
+                ) {
                     $this->fileSystem->delete($newBackgroundImageThumbnail);
                 }
